GROUP6-CROPWHIZ/requirements.txt
absl-py==0.12.0
asgiref==3.3.1
astor==0.8.1
cached-property==1.5.2
certifi==2020.12.5
click==7.1.2
Django==3.1.7
djangorestframework==3.12.4
Flask==1.1.2
gast==0.4.0
google-pasta==0.2.0
grpcio==1.36.1
gunicorn==20.1.0
h5py~=3.9.0
importlib-metadata==3.10.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==1.1.1
numpy~=1.25.2
pandas~=2.1.0
Pillow==8.2.0
protobuf==3.15.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
scikit-learn~=1.3.0
scipy==1.6.2
six==1.15.0
sqlparse==0.4.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
threadpoolctl==2.1.0
typing-extensions==3.7.4.3
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.4.1

matplotlib~=3.7.2
sklearn~=0.0